+++ src/main/java/de/tum/in/naturals/set/BoundedWrapper.java
@@ -17,6 +17,7 @@
 
 package de.tum.in.naturals.set;
 
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import it.unimi.dsi.fastutil.HashCommon;
 import it.unimi.dsi.fastutil.ints.IntCollection;
 import it.unimi.dsi.fastutil.ints.IntIterator;
@@ -139,8 +140,18 @@ class BoundedWrapper extends AbstractBoundedNatBitSet {
   }
 
   @Override
+  @Nonnegative
+  @SuppressFBWarnings(value = "TQ_NEVER_VALUE_USED_WHERE_ALWAYS_REQUIRED",
+                      justification = "Findbugs doesn't infer @Nonnull from control flow")
   public int firstInt() {
-    return complement ? nextPresentIndex(0) : delegate.firstInt();
+    if (complement) {
+      int nextPresent = nextPresentIndex(0);
+      if (nextPresent < 0) {
+        throw new NoSuchElementException();
+      }
+      return nextPresent;
+    }
+    return delegate.firstInt();
   }
 
   @Override
@@ -198,7 +209,7 @@ class BoundedWrapper extends AbstractBoundedNatBitSet {
       if (isEmpty()) {
         throw new NoSuchElementException();
       }
-      return NatBitSets.findLastSetIndex(delegate, domainSize());
+      return NatBitSets.previousPresentIndex(delegate, domainSize() - 1);
     }
     return delegate.lastInt();
   }