+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # 0.x
 
-## 0.2.x
+### 0.3.0 (2017-08-28)
+
+  * Added streams and `clearFrom` method
+  * Fixed wrong behaviour of `LongNatBitSet`'s `clear` method
+  * Added some more functionality to `BitSets` (e.g. `powerSet(BitSet)`) and `NatBitSets` (e.g. `powerSet(NatBitSet)` and `boundedFilledSet(int)`).
+  * Fixed rare bug in `SparseBitSet`
+  * Fixed bug in `NaturalsTransformer`
 
 ### 0.2.0 (2017-08-04)
 
@@ -8,8 +14,6 @@
  * More test cases for `NatBitSet` implementations.
  * Added set variants based on a single `long`.
  * Several utility methods added to `NatBitSets`.
-
-## 0.1.x
 
 ### 0.1.0 (2017-07-23)